Update the model objects to be wrappers around protos. They hold the proto as a private field, and provide accessors which return wrapped objects. No longer have a SourceFile, for now. Also change the AST so that a block is a list of expressions. The distinction between statement and expression is only strong when "if-else" and "while" and so forth have no value.
I'm hoping to merge this branch to trunk before long. This means the grammar will stay broken for a while, but I hope it's justified by being able to make more serious changes in the near future without having to refactor the ANTLR grammar and update source-code-based examples to cope with grammar changes. We can revisit syntax later on when the AST and execution model are more complete.