Dump code to XML as well, to get a sense of how serialization will work, and thanks to encouragement from stevey.
It feels backwards to assemble code out of deltas, especially when I started working on deserializing from XML. It also means that the schema of the code is only expressed in the "adoptChild" methods, which did a bunch of nasty casting to take the untyped relationships in the graph and reassemble a valid tree. Bobby pointed this out in a code review yesterday.

Instead, just assemble the tree the traditional way, by adding children to their parents, using typed methods. Will work on AOP instrumentation of the model classes, as in Hibernate, so that when they're registered as persistant objects, we can record the stream of deltas as the mutator methods are called.